requirements.txt
blinker==1.7.0
Brotli==1.1.0
cachelib==0.10.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
Flask==3.0.0
Flask-Session==0.5.0
fonttools==4.46.0
html5lib==1.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
Pillow==10.1.0
protobuf==4.21.12
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
requests==2.31.0
six==1.16.0
stripe==7.7.0
tinycss2==1.2.1
typing_extensions==4.8.0
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
zopfli==0.2.3
